+++ baggage_setter_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/opentracing/opentracing-go"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/pushself/jaeger-lib/metrics/metricstest"
+	"github.com/uber/jaeger-lib/metrics/metricstest"
 
 	"github.com/pushself/jaeger-client-go/internal/baggage"
 )