+++ config/config.go
@@ -28,7 +28,7 @@ import (
 	throttler "github.com/pushself/jaeger-client-go/internal/throttler/remote"
 	"github.com/pushself/jaeger-client-go/rpcmetrics"
 	"github.com/pushself/jaeger-client-go/transport"
-	"github.com/pushself/jaeger-lib/metrics"
+	"github.com/uber/jaeger-lib/metrics"
 )
 
 const defaultSamplingProbability = 0.001