+++ config/config_test.go
@@ -25,8 +25,8 @@ import (
 	"github.com/opentracing/opentracing-go/ext"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/pushself/jaeger-lib/metrics"
-	"github.com/pushself/jaeger-lib/metrics/metricstest"
+	"github.com/uber/jaeger-lib/metrics"
+	"github.com/uber/jaeger-lib/metrics/metricstest"
 
 	"github.com/pushself/jaeger-client-go"
 	"github.com/pushself/jaeger-client-go/log"