+++ CHANGELOG.md
@@ -280,7 +280,7 @@ Minor patch via https://github.com/jaegertracing/jaeger-client-go/pull/468
 - Support 128bit Trace IDs
 - Change trace/span IDs from uint64 to strong types TraceID and SpanID
 - Add Zipkin HTTP B3 Propagation format support #72
-- Rip out existing metrics and use github.com/pushself/jaeger-lib/metrics
+- Rip out existing metrics and use github.com/uber/jaeger-lib/metrics
 - Change API for tracer, reporter, sampler initialization